ax_python_devel: Add an option to make it optional
ax_python_devel will fail the autoconf run if it doesn't detect that a
python development environment is available.  That's not always what
the user wants, they may want to go on without the functionality.

This change adds a way to pass in an option to specify if you want it to
fail or go on if the functionality is not found.  The default is to
fail, as before.

Signed-off-by: Corey Minyard <minyard@acm.org>