ax_prog_cc_for_build: Fix BUILD_{EXE,OBJ}EXT for Autoconf 2.70
Autoconf 2.70 introduced a change in AC_SUBST behavior that the variable
name in AC_SUBST no longer gets macro expanded. This breaks the rename
syntaxes of BUILD_EXEEXT and BUILD_OBJEXT in AX_PROG_CC_FOR_BUILD.
Change to an alternative approach of assigning BUILD_EXEEXT and
BUILD_OBJEXT manually. The values of EXEEXT and OBJEXT have to be fixed
too at the end of the AX_PROG_CC_FOR_BUILD code.