AX_C_FLOAT_WORDS_BIGENDIAN: fix bug with Emscripten target
Commit 23be7ccd7f306fbc73b3f33e2f5ead716d120eef "Fix AX_C_FLOAT_WORDS_BIGENDIAN
fails whenever interprocedural optimization is enabled." introduced a regression
on the Emscripten target because it switched from grepping the object file where
the float definitely will live, to grepping conftest$EXEEXT which in the case of
Emscripten will be conftest.js but the float is in conftest.wasm. This problem
is unique to Emscripten, it is not present on any other webassembly platform.

To fix, grep `conftest*` for the float.