ax_luarocks_rock.m4: Use precious variable LUA_VERSION if present
It is quite common to have more than one Lua VM available, and
`luarocks` will typically default to one that the system has suggested
as the default. This may or not be the one a developer is trying to
target, especially if LuaJIT is allowable. The AX_PROG_LUA macro now has
detection support for LuaJIT, and this very likely will *not* be the VM
that LuaRocks defaults to looking for.

With this, any previous use of AX_PROG_LUA will have set the variable
LUA_VERSION, and if so `luarocks` will search for manifests matching
that version. This way LuaJIT (which uses 5.1 versioning) can be queried
for compatible rocks even if the system default is 5.4 or similar.