ax_lua_require.m4: add module loader as alternative to lua rock manifest check
The right way to know if a rock is available is check the system (or
user or project) manifest, but this does not always do what it is
supposed to. Many distros do not make the manifest available even when
installing a rock at the system level. The only way to determine if it
is actually available is to try to load it. This check does just that.

Results should be aligned pretty well with AX_LUAROCKS_ROCK, but the
input is a module spec instead of the rockname.