ax_lua_module.m4: combine AX_LUAROCKS_ROCK with fallback to AX_LUA_REQUIRE
This is useful for cases where finding a LuaRock instealled via
`luarocks` or by system package managers is the right way to find it,
but the necessary manifest files may or may not have been properly
installed even if the expected module actually is available. This allows
a fallback to just trying to load the module even if the manifest check
failed and only completely fails if neither method worked.