ax_require_prog.m4: add macro for requiring a prog, not just checking for it
By far one of the most frequest boiler-plate things I do in my projects'
autoconf setup is checking for programs as dependencies. I started using
variations of this macro many years ago and find myself doing something
similar on every single project. This should be a cleaned up version
adapted from various iterations I see across my procjects.

The venerable AC_PATH_PROG macro is great as far as it goes, but it only
handles setting a *value* if the program is not found, not an *action*.
That is easy enough to follow up with a conditional check and custom
error, but the task is very boilerplate and I rarely find myself needing
anything creative, just an error.

Likewise AX_WITH_PROG is nice in that it wraps AC_PATH_PROG with some
user facing --with-EXECUTABLE configure flags and variable handling, bit
it still only handles *values* if not found, not *actions*.

The standard prog checks (such as AC_PROG_SED, AC_PROG_AWK, etc.) do
error if not found, but those only cover specific programs and are not
customizable to adapt to any program a build may need.

Finally the autoconf-archive is full of dozens of custom variants of
AX_PROG_[EXECUTABLE] that error when some executable is not found, but
you need to load the specific macro snippet for each one, and there seem
to be no generic ones that can be used for any program.