ax_openmp: check -lomp when using -nostdlib
when compileing with both "-fopenmp" and "-nostdlib", clang does not automatically
add "-lomp" whereas gcc always add "-lgomp". See below:

With gcc, "-lgomp" is always added:

  # echo 'int main() { return 0; }' | gcc -fopenmp -v -x c - 2>&1 | grep -- "-lgomp"
   /usr/libexec/gcc/x86_64-pc-linux-gnu/14/collect2 -plugin ... -lgomp -lgcc ...
  # echo 'int main() { return 0; }' | gcc -fopenmp -nostdlib -v -x c - 2>&1 | grep -- "-lgomp"
   /usr/libexec/gcc/x86_64-pc-linux-gnu/14/collect2 -plugin ... -lgomp

With clang, "-lomp" is not added if "-nostdlib" is used together with "-fopenmp":

  # echo 'int main() { return 0; }' | clang -fopenmp -v -x c - 2>&1 | grep -- "-lomp"
   "/usr/bin/x86_64-pc-linux-gnu-ld.bfd" --hash-style=gnu ... -lomp -L/usr/lib/llvm/19/lib64 ...
  # echo 'int main() { return 0; }' | clang -fopenmp -nostdlib -v -x c - 2>&1 | grep -- "-lomp"

See the following Gentoo bugs for more context:

> https://bugs.gentoo.org/740310
> https://bugs.gentoo.org/881311
> https://bugs.gentoo.org/896220
> https://bugs.gentoo.org/915833

When libtool add "-nostdlib" but omit "-lomp" when building a shared object (.so), the
resulting .so does not contain "libomp.so" in its dynamic section. For example:

  # LC_ALL=C objdump -x /usr/lib64/libSoundTouch.so
  ...
  Dynamic Section:
    NEEDED               libomp.so

  ...

As a result, any program that depends on the .so may either:
* Fail to build due to unresolved symbols, or
* Fail to start, if the .so has been rebuilt with OpenMP support but lacks proper linkage to libomp.

Signed-off-by: Z. Liu <zhixu.liu@gmail.com>